Moved S3 bucket file to a more generic location and updated test permissions to format properly
