+++ src/functions/pollJournals/framework/databases/mysql/__mocks__/non-test-activites.mock.ts
@@ -1,6 +1,6 @@
 import { ExaminerNonTestActivity } from '../../../../domain/examiner-non-test-activity';
 import { info } from '@dvsa/mes-microservice-common/application/utils/logger';
-import { getMockJournalData } from '../../s3bucket/S3MockJournalsRepository';
+import { getMockJournalData } from '../../../../../../common/framework/s3bucket/S3MockJournalsRepository';
 import { isWithinInterval } from 'date-fns';
 
 export const getMockNonTestActivities = async (
