+++ src/functions/pollJournals/framework/databases/mysql/__mocks__/personal-commitment.mock.ts
@@ -1,6 +1,6 @@
 import { ExaminerPersonalCommitment } from '../../../../domain/examiner-personal-commitment';
 import { info } from '@dvsa/mes-microservice-common/application/utils/logger';
-import { getMockJournalData } from '../../s3bucket/S3MockJournalsRepository';
+import { getMockJournalData } from '../../../../../../common/framework/s3bucket/S3MockJournalsRepository';
 
 export const getMockPersonalCommitments = async (staffNumbers: number[]): Promise<ExaminerPersonalCommitment[]> => {
   let slots: ExaminerPersonalCommitment[] = [];
