+++ src/functions/pollJournals/framework/databases/mysql/__mocks__/test-slot.mock.ts
@@ -1,5 +1,5 @@
 import { ExaminerTestSlot } from '../../../../domain/examiner-test-slot';
-import { getMockJournalData } from '../../s3bucket/S3MockJournalsRepository';
+import { getMockJournalData } from '../../../../../../common/framework/s3bucket/S3MockJournalsRepository';
 import { info } from '@dvsa/mes-microservice-common/application/utils/logger';
 import { isWithinInterval } from 'date-fns';
